bugfix: attributes now taken into account
