+++ Tracer.Fody/NuGet/Tracer.Fody.nuspec
@@ -2,7 +2,7 @@
 <package>
   <metadata>
     <id>Tracer.Fody</id>
-    <version>3.0.0-beta</version>
+    <version>3.0.0-beta2</version>
     <authors>csnemes</authors>
     <owners>csnemes</owners>
     <projectUrl>https://github.com/csnemes/tracer</projectUrl>
